SemanticTool/requirements.txt
pandas
numpy
faiss-cpu
torch
sentence-transformers
